requirements.txt__.txt
discord.py>=2.3.0
pymongo>=4.5.0
dnspython>=2.1.0
python-dotenv>=1.0.0
pillow>=10.0.0
aiohttp>=3.8.0
asyncio>=3.4.3
express>=0.0.0
flask>=2.3.0
requests>=2.31.0
captcha>=0.4.0